refactor: reorganize file service classes into dedicated api, storage, metadata, naming, and util packages
