+++ file/src/main/java/pt/estga/file/services/storage/FileStorageService.java
@@ -1,4 +1,4 @@
-package pt.estga.file.services;
+package pt.estga.file.services.storage;
 
 import org.springframework.core.io.Resource;
