+++ file/src/main/java/pt/estga/file/services/storage/FileStorageServiceLocalImpl.java
@@ -1,4 +1,4 @@
-package pt.estga.file.services.impl;
+package pt.estga.file.services.storage;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -6,7 +6,6 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.core.io.Resource;
 import org.springframework.core.io.UrlResource;
 import org.springframework.stereotype.Service;
-import pt.estga.file.services.FileStorageService;
 import pt.estga.sharedweb.exceptions.FileNotFoundException;
 import pt.estga.sharedweb.exceptions.FileStorageException;
