+++ file/src/main/java/pt/estga/file/services/storage/FileStorageServiceMinioImpl.java
@@ -1,4 +1,4 @@
-package pt.estga.file.services.impl;
+package pt.estga.file.services.storage;
 
 import io.minio.*;
 import io.minio.errors.MinioException;
@@ -8,7 +8,6 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.core.io.InputStreamResource;
 import org.springframework.core.io.Resource;
 import org.springframework.stereotype.Service;
-import pt.estga.file.services.FileStorageService;
 import pt.estga.sharedweb.exceptions.FileNotFoundException;
 import pt.estga.sharedweb.exceptions.FileStorageException;
