+++ file/src/main/java/pt/estga/file/util/CountingInputStream.java
@@ -1,4 +1,4 @@
-package pt.estga.file.services;
+package pt.estga.file.util;
 
 import lombok.Getter;
 import org.jspecify.annotations.NonNull;
