+++ mark/src/main/java/pt/estga/mark/listeners/MarkDetectionListener.java
@@ -14,7 +14,7 @@ import pt.estga.mark.repositories.MarkOccurrenceRepository;
 import pt.estga.mark.repositories.MarkRepository;
 import pt.estga.detection.DetectionResult;
 import pt.estga.detection.DetectionService;
-import pt.estga.file.services.MediaService;
+import pt.estga.file.services.api.MediaService;
 
 import java.io.IOException;
 import java.util.Optional;
