+++ mark/src/main/java/pt/estga/mark/services/MarkOccurrenceService.java
@@ -10,7 +10,7 @@ import pt.estga.mark.entities.MarkOccurrence;
 import pt.estga.mark.events.MarkOccurrenceCreatedEvent;
 import pt.estga.mark.repositories.MarkOccurrenceRepository;
 import pt.estga.file.entities.MediaFile;
-import pt.estga.file.services.MediaService;
+import pt.estga.file.services.api.MediaService;
 
 import java.io.IOException;
 import java.util.Optional;
